+++ tests/CSVArrayTest.php
@@ -41,7 +41,11 @@ final class CSVArrayTest extends TestCase
         ];
     }
 
-    /** @param array<int, array<string, string>> $array */
+    /**
+     * @dataProvider csvAndArrayStringValues
+     *
+     * @param array<int, array<string, string>> $array
+     */
     #[\PHPUnit\Framework\Attributes\DataProvider('csvAndArrayStringValues')]
     public function testStringValues(string $csv, array $array): void
     {