+++ tests/Microplate/CoordinateSystemTest.php
@@ -11,6 +11,7 @@ use PHPUnit\Framework\TestCase;
 final class CoordinateSystemTest extends TestCase
 {
     #[\PHPUnit\Framework\Attributes\DataProvider('firstLast')]
+    /** @dataProvider firstLast */
     public function testFirstLast(CoordinateSystem $coordinateSystem, string $expectedFirst, string $expectedLast): void
     {
         $actualFirst = $coordinateSystem->first();